refactor partioning bricks into partition diretory
